refactor(dm-tool): split compendium projection into per-type files
Split the 1,075-line monolithic `projection.ts` into a
`projection/` directory with four focused modules:

- `shared.ts`  — defensive readers, cleanDescription, portrait/token URL helpers
- `monster.ts` — MonsterRow/MonsterResult types, all monster formatters and projections
- `item.ts`    — price helpers and all item projections
- `index.ts`   — barrel re-export matching the previous public surface exactly

`prepared.ts` updated to import from `./projection/index.js`.
Existing `projection.test.ts` (801 lines) split into:
- `projection/monster.test.ts` — monster + shared-formatter coverage
- `projection/item.test.ts`    — item + price-helper coverage

Zero logic changes; all 405 tests pass.
